seat: Move video device handling to the seat.
